Fix: Stylecode, create a separate file for coredump
