+++ libr/debug/p/debug_native.c
@@ -1276,8 +1276,7 @@ static bool r_debug_gcore (RDebug *dbg, RBuffer *dest) {
 #if __APPLE__
 	return xnu_generate_corefile (dbg, dest);
 #elif __linux__ && __x86_64__
-	r_sandbox_disable (true);
-	return linux_generate_corefile(dbg, dest);
+	return linux_generate_corefile (dbg, dest);
 #else
 	return false;
 #endif
